fix(deps): upgrade postgres version for 13 to 13.12.0
Version 13.11 is vulnerable to CVE-2023-39417, which exists in versions >= 13.0, < 13.12.

The vulnerability was found in the National Vulnerability Database (NVD) based on the CPE cpe:2.3:a:postgresql:postgresql with NVD severity: High.

The file is associated with the technology PostgreSQL.

The vulnerability can be remediated by updating PostgreSQL to 13.12 or higher.